moving profile permission dashboard links to separate component
